bug-fix + refactoring (Req-13)
implemented helper class "Pair".
Renamed and moved GitCommitMessageParser...
